+++ src/main/java/com/techfork/global/security/handler/login/OAuth2AuthenticationSuccessHandler.java
@@ -6,7 +6,7 @@ import com.techfork.global.security.jwt.JwtDTO;
 import com.techfork.global.security.jwt.JwtProperties;
 import com.techfork.global.security.jwt.JwtUtil;
 import com.techfork.global.security.oauth.UserPrincipal;
-import com.techfork.global.util.CookieUtil;
+import com.techfork.global.security.util.CookieUtil;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
